created validators folder, moved form-text area to validators folder
